update uninstall and utils
